requirements.txt
praw>=7.7.0
python-dotenv>=1.0.0
click>=8.1.0
streamlit>=1.30.0
google-genai>=1.0.0
